Moved styles to styles directory. Added base style folder and moved some base styles there.
