+++ src/Components/zones/EditZone.tsx
@@ -1,5 +1,4 @@
 import { Box, Modal, TextField, Typography } from "@mui/material";
-import "./AddZone.css";
 import Button from "@mui/material/Button";
 import agent from "../../App/api/agent";
 import { Formik, Form, Field, ErrorMessage } from "formik";
@@ -7,6 +6,7 @@ import * as Yup from "yup";
 import { useSelector } from "react-redux";
 import { RootState } from "../../redux/store";
 import { useEffect } from "react";
+import "../../styles/zones/AddZone.css";
 
 type ZoneBarProps = {
   fetchZones(args: string): void;
