+++ src/styles/baseStyles/BaseCard.css
@@ -1,3 +1,8 @@
+.card-content-wrapper {
+  padding: 16px 0 0 0 !important;
+  border-radius: 0 0 15px 15px !important;
+}
+
 .card-img {
   position: relative;
 }
@@ -51,7 +56,7 @@
   transform: scale(0.8);
 }
 
-.zone-name {
+.card-name {
   color: #666666;
   font-weight: 700 !important;
   font-family: "Outfit", sans-serif !important;
@@ -81,37 +86,6 @@
 .card-data span:nth-child(2) {
   padding-left: 0.5rem !important;
   font-weight: 200;
-}
-
-.card-btn {
-  font-size: 1.5rem !important;
-  width: 100% !important;
-  min-width: 50px !important;
-  background-color: #ffffffa8 !important;
-  color: #002b49 !important;
-  border-radius: 20px !important;
-  padding: 0 !important;
-}
-
-#card-details {
-  width: 3.75rem !important;
-}
-
-#card-details .action-icon {
-  font-size: 1.75rem !important;
-}
-
-#card-delete .action-icon {
-  font-size: 1.25rem !important;
-}
-
-.card-btn:hover {
-  background-color: #ffffffe8 !important;
-}
-
-.card-zone-data {
-  padding: 16px 0 0 0 !important;
-  border-radius: 0 0 15px 15px !important;
 }
 
 .gallons-chip {
