+++ src/styles/plants/PlantBar.css
