+++ src/styles/plants/ViewPlant.css
